+++ Mods ChoGGi/Get More Tourists/metadata.lua
@@ -3,8 +3,8 @@ return PlaceObj("ModDef", {
 		PlaceObj("ModDependency", {
 			"id", "ChoGGi_Library",
 			"title", "ChoGGi's Library",
-			"version_major", 11,
-			"version_minor", 9,
+			"version_major", 12,
+			"version_minor", 0,
 		}),
 	},
 	"title", "Get More Tourists",